Begin implementing the localdocs ui in earnest.
